没有用线程同步还更稳定？ 以前使用线程同步，收到的数据滞后，lastCmd 超前，使得返回数据和命令对应不上。

+++ src/main/java/com/rytec/rec/node/DAM/DAMOutput.java
@@ -108,7 +108,7 @@ public class DAMOutput extends NodeOutput implements NodeInterface {
     //消息解码
     public NodeMessage decodeMessage(ChannelMessage msg) {
 
-        int value = 0;
+        byte value = 0;
         ChannelMessage respond = msg;
 
         switch (respond.type) {
